TST: split up test_concat.py #37243
